Add files via upload
I use Quanteda package in R to conduct corpus, trim data, and run the LDA model, and to replicate as much as I could do, I apply the threshold and stopwords list from Amy Catalinac.  Also, I set the topic number to 69 in order to get a similar result. I found that the words each topic contained are highly similar to Amy's result, but only in Japanese since I am not pretty familiar with Japanese and I'll translate the result maybe by DeepL or other translator.